Remove @bszwarc and @tomekpapiernik from CODEOWNERS
Basia and Tomek left Kyma in September and have not continued contributing to our project since then. According to the [offboarding guidelines](https://kyma-project.io/community/governance#kyma-working-model-kyma-working-model-when-does-a-maintainer-lose-the-maintainer-status), a person should be removed from codeowners in case they are no longer interested in contributing or haven't contributed to the project for more than 3 months. 

Changes proposed in this pull request:

- Remove @bszwarc from CODEOWNERS