Add override interceptor for container registry (#276)
We want to use k3d's container registry for integration tests. This
change adds an override interceptor to set the registry address for the
serverless component to k3d registry.

Using the k3d registry for serverless function builds requires also a
change to the CoreDNS configuration, which will be added in a separate
PR.